xen-dom-mgmt: dom0: add dom0 initcall routine
During Zephyr xenlib initialization some actions (root node init in
Xenstore, Domain-0 Xenstore entries adding) should be done right
before application start. Previously these actions were implemented
in separate initcalls with different priorities which is a little bit
complicated.

As we will need to have Domain-0 struct in runtime soon (for further
implementation of xss watch functions in Xenstore API), Domain-0
related actions were moved to single routine: root node init was made
global and xenstore entries adding was refatored. So now we can
simplify priorities for initcalls and have a single place where all
pre-application actions can be done.

Acked-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Acked-by: Dmytro Semenets <dmytro_semenets@epam.com>
